staging: fsl-dpaa2: Account extra headroom in skb buffer
Seeding the buffer pools is done with an extra headroom, to prevent
reallocations on the IP forwarding path. This extra space is further
used to be a part of the skb buffer headroom, but the current code
does not make it a part of the total skb alloc'd size. In scenarios
with MTU size close to the Rx buffer size, the few bytes eaten by
the extra skb headroom might trigger a skb_panic.

Account the extra headroom when allocating the skb.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>