staging: fsl-dpaa2: Add pause frame support
Add L2 flow control support, configurable via ethtool.

This has an impact on the number of buffers available for a NI:
- If pause frames are enabled, there's a small number of available
  buffers in the pool. Testing shows that using a large number of
  buffers doesn't improve forwarding results. In forwarding scenarios
  with a large number of buffers, there's a tendency for a single core
  to fill up all available buffers with pending frames and never finish
  the work (due to memory access latencies / cache misses), so it's
  better to keep the number of buffers small for better latencies.
- If pause frames are disabled, more buffers are allocated and taildrop
  thresholds are enabled per Rx queue.

Pause frames can't be enabled / disabled if the interface is up and
autonegotiation is on.

By default, pause frames are disabled for an NI.

Move taildrop configuration to a separate function.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>