staging: fsl-dpaa2/eth: Use 64B alignment for Rx buffer
Reduce the enforced Rx alignment from 256B to 64B. The old
value was related to a LS2080 hw limitation, but it wasn't
actually required in order to avoid that errata. So fall back
to the default 64B hardware requirement for buffer alignment.

Also fix an issue where changing the Rx align value forced
the stack to realloc the skb in forwarding scenarios.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>