staging: fsl-dpaa2: Update buffer count for flow control
The DPAA2 Ethernet driver sets the number of buffers to seed the pool on
a per queue basis. With the current code, network interfaces that have a
small number of Rx queues will trigger the flow control depletion
threshold, causing pause frames to be sent continuously and stopping
traffic.

In order to account for these scenarios, when flow control pause frames
are enabled, use a per buffer pool count, which is distributed to the
number of channels.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>